relocate swagger config to config carpet
