--- config/model.txt
@@ -1,12 +1,0 @@
-// "components": 
-// {
-// 	"schemas": 
-// 	{
-// 		"question": 
-// 		{
-// 			"properties": {
-// 				"question": { "type": "string", "required": true}, "codeExamples": { "type": [ "string" ] }, "answerOptions": [{ "answer": { "type": "string", "required": true }, "isCorrect": { "type": "boolean", "required": true }}],
-// 			},
-// 		},
-// 	},
-// }
