+++ routes/api.routes.js
@@ -20,6 +20,6 @@ router.get("/v1/questions/random", apiControllers.getRandomQuestions);
 
 router.get("/v1/questions/ai", customLimiter, apiControllers.getAiQuestions);
 
-router.get('/v1/test-swagger', apiControllers.getTestSwaggerOption);
+router.get('/v1/swagger-docs', apiControllers.getSwaggerDocs);
 
 module.exports = router;
